% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_a_basic_info.R
\name{block_a_basic_info}
\alias{block_a_basic_info}
\title{Update Block A of SELECTOR.IN: Basic Information}
\usage{
block_a_basic_info(hydrus_model)
}
\arguments{
\item{hydrus_model}{a hydrus model created with \code{\link{create_hydrus_project}}}
}
\value{
Edits BLOCK A of SELECTOR.IN file associated with the input hydrus_model
}
\description{
Update Block A of SELECTOR.IN: Basic Information
}
\examples{
block_a_basic_info(hydrus_model)

}
