% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_b_water_flow_info.R
\name{block_b_water_flow_info}
\alias{block_b_water_flow_info}
\title{Set Block B of SELECTOR.IN: Water Flow Information}
\usage{
block_b_water_flow_info(hydrus_model)
}
\arguments{
\item{hydrus_model}{a hydrus model created with \code{\link{create_hydrus_project}}}
}
\value{
edits SELECTOR.IN file
}
\description{
Set Block B of SELECTOR.IN: Water Flow Information
}
\examples{
block_b_water_flow_info(hydrus_model)

}
