% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_c_time_info.R
\name{block_c_time_info}
\alias{block_c_time_info}
\title{Set Time Parameters in BLOCK C of SELECTOR.IN}
\usage{
block_c_time_info(hydrus_model)
}
\arguments{
\item{hydrus_model}{a hydrus model created with \code{\link{create_hydrus_project}}}
}
\value{
edits BLOCK C of SELECTOR.IN
}
\description{
Set Time Parameters in BLOCK C of SELECTOR.IN
}
\examples{
block_c_time_info(hydrus_model)
}
