% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_d_root_growth_info.R
\name{block_d_root_growth_info}
\alias{block_d_root_growth_info}
\title{Set Block D of SELECTOR.IN: Root Growth Information}
\usage{
block_d_root_growth_info(hydrus_model)
}
\arguments{
\item{hydrus_model}{a hydrus model created with \code{\link{create_hydrus_project}}}
}
\value{
edits BLOCK D of SELECTOR.IN
}
\description{
Set Block D of SELECTOR.IN: Root Growth Information
}
\examples{
block_d_root_growth_info(hydrus_model)
}
