% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_f_solute_transport_info.R
\name{block_f_solute_transport_info}
\alias{block_f_solute_transport_info}
\title{Set Block F of SELECTOR.IN: Solute Transport Information}
\usage{
block_f_solute_transport_info(hydrus_model)
}
\arguments{
\item{hydrus_model}{a hydrus model created with \code{\link{create_hydrus_project}}}
}
\value{
edits SELECTOR.IN file
}
\description{
Set Block F of SELECTOR.IN: Solute Transport Information
}
\examples{
block_f_solute_transport_info(hydrus_model)

}
