% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_g_root_water_uptake_info.R
\name{block_g_root_water_uptake_info}
\alias{block_g_root_water_uptake_info}
\title{Set Block G of SELECTOR.IN: Root Water Uptake Information}
\usage{
block_g_root_water_uptake_info(hydrus_model)
}
\arguments{
\item{hydrus_model}{a hydrus model created with \code{\link{create_hydrus_project}}}
}
\value{
edits BLOCK G of SELECTOR.IN
}
\description{
Set Block G of SELECTOR.IN: Root Water Uptake Information
}
\examples{
block_g_root_water_uptake_info(hydrus_model, root_water_uptake)
}
