% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_for_model_errors.R
\name{check_for_model_errors}
\alias{check_for_model_errors}
\title{HYDRUS model logic check}
\usage{
check_for_model_errors(hydrus_model)
}
\arguments{
\item{hydrus_model}{}
}
\value{
returns nothing if no errors, will stop function and output error if there is any in the HYDRUS model parameterization
}
\description{
HYDRUS model logic check
}
\examples{
check_for_model_errors(hydrus_model)
}
