% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_profile_modelinputs.R
\name{create_profile_modelinputs}
\alias{create_profile_modelinputs}
\title{Title}
\usage{
create_profile_modelinputs(
  Zf,
  obs9,
  material.number,
  material.divide,
  model.number
)
}
\arguments{
\item{model.number}{}
}
\description{
Title
}
