% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_selector_in.R
\name{create_selector_in}
\alias{create_selector_in}
\title{Create and edit SELECTOR.IN file}
\usage{
create_selector_in(
  hydrus_model,
  time_var_bc = FALSE,
  meteorological_data = FALSE,
  internal_interpolation_tables = TRUE,
  internal_interp_params = list(lower_limit_tension_interval = 1e-06,
    upper_limit_tension_interval = 10000),
  soil_hydraulic_model = 0,
  hysteresis = 0,
  initially_drying_curve = TRUE,
  soil_hydraulic_parameters = data.frame(material = 1, theta_r = 0.045, theta_s = 0.43,
    alpha = 0.145, n = 2.68, K_s = 712.8, l = 0.5),
  iteration_criteria = list(maximum_iterations = 10, water_content_tol = 0.001,
    pressure_head_tol = 1),
  water_flow_bcs = list(upper_bc = "constant_pressure_head", lower_bc = "free_drainage",
    triggered_irrigation = FALSE),
  particle_tracking_params = list(init_water_storage = 2, cumulative_surface_flux = 2),
  time_parameters = list(initial_time_step = 0.001, minimum_time_step = 1e-05,
    maximum_time_step = 5, lower_time_step_mult = 1.3, upper_time_step_mult = 0.7,
    lower_optim_iter_range = 3, upper_optim_iter_range = 7, initial_model_time = 0,
    final_model_time = 100)
)
}
\arguments{
\item{hydrus_model}{a hydrus model object created with
create_hydrus_project()}

\item{time_var_bc}{logical. Use time variable boundary conditions (TRUE) or
not (FALSE).}

\item{meteorological_data}{logical. Option to use meteorological data in time
variable boundary condition?}

\item{soil_hydraulic_model}{integer indicating which soil hydraulic model to
use. See details.}

\item{hysteresis}{integer indicating the type of hysteresis in a water flow
model using van Genchten-Maulem hydraulics.}

\item{initially_drying_curve}{logical. Is the hysteresis initially drying
(TRUE) or initially wetting curve (FALSE)?}

\item{soil_hydraulic_parameters}{data.frame indicating the proper parameters
needed for the soil hydraulic model. Parameter values for each material
separated by row.}

\item{inverse}{logical. Regular, forward model is FALSE, inverse model is
TRUE}

\item{snow_hydrology}{logical. Snow hydrology option of water flow model.
Heat transport must also be turned on.}

\item{vapor_flow}{logical. Vapor flow option of water flow model.}

\item{particle_tracking}{logical. Particle tracking option of water flow
model.}

\item{solute_transport_model}{list of solute transport model options}

\item{root_water_uptake_opts}{list of root water uptake options}

\item{print_options}{list of print options.}
}
\description{
Further set model parameters that are edited and set in the SELECTOR.IN file
of the HYDRUS project. Some model parameters are taken from the hydrus model
object creates using 'create_hydrus_project' and many others are set within
this function. the hydrus model object is edited to reflect the newly set
parameters.
}
\details{
The 'soil_hydraulic_model' options are: Single-porosity van Genuchten-Maulem
= 0 Single-porosity van Genuchten-Maulem with air-entry value of -2 cm = 3
Single-porosity modified van Genuchten (Vogel and Cislerova 1988) = 1
Single-porosity Brooks-Corey (1964) = 2 Kosugi (1996) (the log normal model)
= 4 Kosugi (1996) (the log normal model) with Brunswick model modification
(Weber et al 2019) = -4 Dual-porosity (Durner, 1994; the dual van
Genuchten-Maulem model) = 5 Dual-porosity (mobile-immobile water, water
content mass transfer) = 6 Dual-porosity (mobile-immobile water, pressure
head mass transfer) = 7 Dual-permeability (add-on Module) = 8 Look-up Table
with Hydraulic capacity calculated or provided = 9

Hysteresis options (only available for the single-porosity van
Genchten-Maulem model with or without air entry value of -2 cm): No
hysteresis = 0 (default) Hysteresis in retention curve (Kool & Parker, 1987)
= 1 Hysteresis in retention curve and conductivity = 2 Hysteresis in
retention curve (no pumping; Lenhard and Parker, 1992) = 3

When hysteresis is turned on, user must specify if the curve is initially
drying ('initially_drying_curve = TRUE') or initially wetting
('initially_drying_curve = FALSE')

'water_flow_boundary_condition' options for 'upper_bc' include:
"constant_pressure_head" (default), "constant_flux",
"atm_bc_with_surface_layer", "atm_bc_with_surface_runoff",
"variable_pressure_head", "variable_pressure_head/flux".
'water_flow_boundary_condition' options for 'lower_bc' include:
"constant_pressure_head", "constant_flux", "variable_pressure_head",
"variable_flux", "free_drainage" (default), "deep_drainage", "seepage_face",
"horizontal_drains".
}
