% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_atmosph_file.R
\name{edit_atmosph_file}
\alias{edit_atmosph_file}
\title{Edit the ATMPOSH.IN file of a HYDRUS model}
\usage{
edit_atmosph_file(
 hydrus_model,
 atm_time_series = data.frame(time = 1,
                              precip = 0,
                              evap = 0,
                              transpiration = 0,
                              min_pressure_head = 100000,
                              top_conc = 0,
                              bot_conc = 0),
 max_h_at_surface = 0
)
}
\arguments{
\item{hydrus_model}{a hydrus_model}

\item{atm_time_series}{data.frame of time-variable atmospheric data}

\item{max_h_at_surface}{maximum head at soil surface}
}
\value{
edits the ATMOSPH.IN file
}
\description{
Edit the ATMPOSH.IN file of a HYDRUS model
}
\examples{
edit_atmosph_file(hydrus_model, max_h_at_surface = 5)
}
