% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_dat_file.R
\name{edit_dat_file}
\alias{edit_dat_file}
\title{Title}
\usage{
edit_dat_file(hydrus_model)
}
\arguments{
\item{hydrus_project}{a hydrus project object created with
'create_hydrus_project'}
}
\value{
Edits the .DAT file of the HYRDUS project
}
\description{
Title
}
\examples{
edit_dat_file(hydrus_project, main_processes = main_processes, model_units = model_units, geometry = geometry, print_options = print_options, water_flow_bcs = water_flow_bcs)
}
