% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_meteo_file.R
\name{edit_meteo_file}
\alias{edit_meteo_file}
\title{Edit the METEO.IN file}
\usage{
edit_meteo_file(
 radiation_type = "potential_radiation", # solar_radiation, net_radiation
 et_equation = "penman-monteith", # hargreaves
 cloudiness_option = "sunshine", #cloudiness, transmission_c, solar_radiation
 penman_montieth_option = "relative_humidity", # vapor_pressure
 crop_option = "no_crop", # constant, tables, daily
 latitude = 40,
 altitude = 110,
 sw_angstrom_a = 0.25,
 sw_angstrom_b = 0.5,
 sw_cloudiness_a = 1.35,
 sw_cloudiness_b = -0.35,
 lw_cloudiness_a = 0.9,
 lw_cloudiness_b = 0.1,
 lw_emissivity_a = 0.34,
 lw_emissivity_b = -0.139,
 windspeed_meas_height = 200,
 temp_meas_height = 200,
 albedo = 0.23,
 meteo_time_series = data.frame(time = 1,
                                radiation = 300,
                                temp_max = 20,
                                temp_min = 10,
                                rh_mean = 95,
                                windspeed = 30,
                                sun_hours = 4,
                                crop_height = 0,
                                albedo = 0,
                                lai = 0,
                                root_depth = 0)
)
}
\arguments{
\item{hydrus_model}{a hydrul_model}

\item{radiation_type}{character string indicating radiation type. See details.}

\item{et_equation}{character string indicating evapotranspiration equation. See details.}

\item{cloudiness_option}{character string indicating the cloudiness option. See detail.}

\item{penman_montieth_option}{character string indicating the Penman-Montieth calculation option. See details.}

\item{crop_option}{character sting indicating how to input crop data. See details.}

\item{latitude}{numeric}

\item{altitude}{numeric}

\item{sw_angstrom_a}{numeric}

\item{sw_angstrom_b}{numeric}

\item{sw_cloudiness_a}{numeric}

\item{sw_cloudiness_b}{numeric}

\item{lw_cloudiness_a}{numeric}

\item{lw_cloudiness_b}{numeric}

\item{lw_emissivity_a}{numeric}

\item{lw_emissivity_b}{numeric}

\item{windspeed_meas_height}{numeric}

\item{temp_meas_height}{numeric}

\item{albedo}{numeric}

\item{meteo_time_series}{data frame of meteorologival data}
}
\value{
edits METEO.IN file
}
\description{
Edit the METEO.IN file
}
\examples{
edit_meteo_file(hydrus_model, cloudiness_option = "cloudiness")
}
