% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizing_functions.R
\name{geometry}
\alias{geometry}
\title{Set units in a HYDRUS model}
\usage{
geometry(
 number_materials = 1,
 number_subregions = 0,
 number_nodes = 101,
 profile_depth = 100,
 observation_nodes_n = 0,
 observation_nodes = NA
 relative_gravity = 1.0
)
}
\arguments{
\item{number_materials}{integer. The number of materials in model.}

\item{number_subregions}{integer. The number of subregions in model.}

\item{number_nodes}{integer. The number of model nodes.}

\item{profile_depth}{integer. Depth of the model soil profile.}

\item{observation_nodes_n}{integer. The number of observation nodes}

\item{observation_nodes}{integer or vector of integers indicating the node number of observation nodes.}

\item{relative_gravity}{numeric. Gravity for Earth vertical = 1; Earth horizontal = 0; Moon = 0.166; Mars = 0.38.}
}
\value{
list of parameters
}
\description{
Function to set model_units while maintaining unused default model parameters.
}
\examples{
hydrus_model$geometry <- geometry(number_materials = 4,
                                            observation_nodes_n = 10,
                                            observation_nodes = c(2, 5, 9, 14, 18, 27, 41, 52, 64, 84))
}
