% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrus_sci_format.R
\name{hydrus_sci_format}
\alias{hydrus_sci_format}
\title{Formats numbers into HYDRUS-specific scientific format}
\usage{
hydrus_sci_format(df)
}
\arguments{
\item{df}{a data.frame of numerics}
}
\value{
a data.frame with the numerics converted to the correct scientific formatting as a character string
}
\description{
Formats numbers into HYDRUS-specific scientific format
}
\examples{
hydrus_sci_format(data.frame(x = 175.0))
}
