% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizing_functions.R
\name{iteration_criteria}
\alias{iteration_criteria}
\title{Set iteration criteria parameters in a HYDRUS model}
\usage{
iteration_criteria(
 maximum_iterations = 10,
 water_content_tol = 0.001,
 pressure_head_tol = 1,
 internal_interpolation_tables = TRUE,
 lower_limit_tension_interval = 1e-06,
 upper_limit_tension_interval = 10000
)
}
\arguments{
\item{maximum_iterations}{integer.}

\item{water_content_tol}{numeric. water content tolerance.}

\item{pressure_head_tol}{numeric. pressure heat tolerance in model space_unit}

\item{internal_interpolation_tables}{logical. Enable internal interpolation tables}

\item{lower_limit_tension_interval}{numeric. Lower limit of the tension interval in space_unit}

\item{upper_limit_tension_interval}{numeric. Upper limit of the tension interval in space_unit}
}
\value{
list of parameters
}
\description{
Function to set iteration_criteria while maintaining unused default model parameters.
}
\examples{
hydrus_model$iteration_criteria <- iteration_criteria(maximum_iterations = 25)
}
