% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizing_functions.R
\name{main_processes}
\alias{main_processes}
\title{Set main processes in a HYDRUS model}
\usage{
main_processes(
 water_flow = TRUE,
 vapor_flow = FALSE,
 snow_hydrology = FALSE,
 particle_tracking = FALSE,
 solute_transport = FALSE,
 unsatchem = FALSE,
 HP1 = FALSE,
 heat_transport = FALSE,
 root_water_uptake = FALSE,
 root_growth = FALSE,
 inverse = FALSE
)
}
\arguments{
\item{water_flow}{logical; if TRUE, model simulates water flow process. Default TRUE.}

\item{vapor_flow}{logical; if TRUE, model simulates vapor flow process. Default FALSE.}

\item{snow_hydrology}{logical; if TRUE, model uses snow hydrology process. Default FALSE. water_flow and heat_transport must both be TRUE.}

\item{particle_tracking}{logical; if TRUE, model simulates particle tracking. Default FALSE.}

\item{solute_transport}{logical; if TRUE, model simulates solute transport process. Default FALSE.}

\item{unsatchem}{logical; if TRUE, model uses UNSATCHEM add-on module. Default FALSE.}

\item{HP1}{logical; if TRUE, model uses HP1 add-on module. Default FALSE.}

\item{heat_transport}{logical; if TRUE, model simulates heat transport processes. Default FALSE.}

\item{root_water_uptake}{logical; if TRUE, model simulates root water uptake process. Default FALSE.}

\item{root_growth}{logical; if TRUE, model simulates root growth process. Default FALSE.}

\item{inverse}{logical; if TRUE, model does an inverse simulation. Default FALSE.}
}
\value{
list of parameters
}
\description{
Function to set main_processes while maintaining unused default model parameters.
}
\examples{
hydrus_model$main_processes <- main_processes(water_flow = TRUE,
                                                        heat_transport = TRUE,
                                                        snow_hydrology = TRUE)

}
