% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizing_functions.R
\name{model_units}
\alias{model_units}
\title{Set units in a HYDRUS model}
\usage{
model_units(
 time_unit = "days",
 space_unit = "cm",
 mass_unit = "mmol"
)
}
\arguments{
\item{time_unit}{character string indicating temporal unit. Must be "seconds", "minutes", "hours", "days" or "years"}

\item{space_unit}{character string indicating the spatial unit. Must be "mm", "cm", or "m".}

\item{mass_unit}{character string indicating the solute mass unit. Options include:}
}
\value{
list of parameters
}
\description{
Function to set model_units while maintaining unused default model parameters.
}
\examples{
hydrus_model$model_units <- model_units(time_unit = "seconds",
                                                  space_unit = "mm")
}
