% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterize_hydrus_model.R
\name{parameterize_hydrus_model}
\alias{parameterize_hydrus_model}
\title{Parameterize a HYDRUS model}
\usage{
parameterize_hydrus_model(hydrus_model)
}
\arguments{
\item{hydrus_model}{a hydrus model created with \code{\link{create_hydrus_project}}}
}
\value{
Function return a list of a model parameters. Edit parameters with \code{\link{main_processes}}, \code{\link{model_units}}, \code{\link{geometry}}, \code{\link{print_options}},
\code{\link{time_parameters}}, \code{\link{time_variable_bc}}, \code{\link{iteration_criteria}}, \code{\link{soil_hydraulics}},
\code{\link{water_flow_bcs}}, \code{\link{root_water_uptake}}, \code{\link{root_growth}}, \code{\link{solute_options}}, and \code{\link{particle_tracking}}.
}
\description{
Parameterize a HYDRUS model
}
\examples{
parameterize_hydrus_model(hydrus_model)

}
