% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizing_functions.R
\name{particle_tracking}
\alias{particle_tracking}
\title{Set particle tracking parameters in a HYDRUS model}
\usage{
particle_tracking(init_water_storage = 2, cumulative_surface_flux = 2)
}
\arguments{
\item{init_water_storage}{numeric. Initial water storage.}

\item{cumulative_surface_flux}{numeric. Cumulative surface flux}
}
\value{
list a parameters
}
\description{
Function to set particle_tracking while maintaining unused default model parameters.
}
\examples{
hydrus_model$particle_tracking <- particle_tracking(init_water_storage = 1)
}
