% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.hydrus_model.R
\name{print.hydrus_model}
\alias{print.hydrus_model}
\title{print.hydrus_model}
\usage{
\method{print}{hydrus_model}(x, indent = 0, color = TRUE)
}
\arguments{
\item{x}{object of class 'hydrus_model'}

\item{indent}{indentation width}

\item{color}{logical, default TRUE}
}
\value{
prints list in pretty way
}
\description{
print.hydrus_model
}
\examples{
print
}
