% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizing_functions.R
\name{print_options}
\alias{print_options}
\title{Set print options in a HYDRUS model}
\usage{
print_options(
 print_times = TRUE,
 screen_output = TRUE,
 print_fluxes_not_temp = TRUE,
 times_to_print = data.frame(times = c(100)),
 interval_output_option = FALSE,
 interval_output = 0.001,
 print_time_information = TRUE,
 time_info_print_every_n_time_steps = 1
)
}
\arguments{
\item{print_times}{logical. Print times (TRUE) on not (FALSE)}

\item{screen_output}{logical. Print output to screen (TRUE) or not (FALSE)}

\item{print_fluxes_not_temp}{logical. Print fluxes (TRUE) or temperature (FALSE)}

\item{times_to_print}{data.frame of times to print in time_units.}

\item{interval_output_option}{logical. Use interval output option (TRUE)}

\item{interval_output}{numeric. The interval to use when the interval output option is used.}

\item{print_time_information}{logical. Print time information (TRUE) or not (FALSE)}

\item{time_info_print_every_n_time_steps}{integer. T-level information every n time steps.}
}
\value{
list of parameters
}
\description{
Function to set print_options while maintaining unused default model parameters.
}
\examples{
hydrus_model$print_options <- print_options(interval_output_option = TRUE,
                                                      interval_output = 5)
}
