% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizing_functions.R
\name{root_growth}
\alias{root_growth}
\title{Set root growth parameters in a HYDRUS model}
\usage{
root_growth(
 root_growth_depth = 0,
 root_growth_params = NA,
 root_growth_factor = NA
)
}
\arguments{
\item{root_growth_depth}{integer indicating how root depth will be specified. See details for options.}

\item{root_growth_params}{a data.frame of root growth parameters for the given root depth specification and root growth factor option}

\item{root_growth_factor}{integer indicating how to specify the root growth factor when using a logisitc growth function to specufy root growth depth (\code{root_growth_depth = 2}). See details for options.}
}
\value{
a list of parameters
}
\description{
Function to set root_growth while maintaining unused default model parameters.
}
\details{
\code{root_growth_depth} options:
\describe{
\item{\code{root_growth_depth = 0}}{Root depth specified with time-variable boundary conditions.}
\item{\code{root_growth_depth = 1}}{Root depth specified with a table. \code{root_growth_params}: "Time", "RootDepth"}
\item{\code{root_growth_depth = 2}}{Root depth specified using a logistic growth function.
   \code{root_growth_factor} options:
   \describe{
   \item{\code{root_growth_factor = 0}}{Root growth factor from given data. \code{root_growth_params}: "initial_root_growth_time", "time_root_data", "harvest_time", "initial_rooting_depth",
   "depth_root_data", "maximum_rooting_depth", "time_period".}
   \item{\code{root_growth_factor = 1}}{Root growth factor 50% after 50% growing season. \code{root_growth_params}: "initial_root_growth_time",
   "harvest_time", "initial_rooting_depth",
   "maximum_rooting_depth", "time_period".}
   }
 }
}
}
\examples{
hydrus_model$root_growth <- root_growth(root_growth_depth = 2,
                                                  root_growth_params = data.frame(initial_root_growth = 115,
                                                  harvest_time = 245,
                                                  initial_rooting_depth = 0,
                                                  maximum_rooting_depth = 60,
                                                  time_period = 365
                                                  ))
}
