% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizing_functions.R
\name{root_water_uptake}
\alias{root_water_uptake}
\title{Set root water uptake parameters in a HYDRUS model}
\usage{
root_water_uptake(
 root_water_uptake_model = 0,
 root_water_uptake_params = data.frame(critical_stress_index = 1,
                                       P0 = -10,
                                       POpt = -25,
                                       P2H = -200,
                                       P2L = -800,
                                       P3 = -8000,
                                       r2H = 0.5,
                                       r2L = 0.1),
 active_solute_uptake = FALSE
)
}
\arguments{
\item{root_water_uptake_model}{integer indicating which root water uptake model to use. See details for options.}

\item{root_water_uptake_params}{data frame of root water uptake parameters. See details for parameters for each model.}

\item{active_solute_uptake}{logical. Turn on active solute uptake (TRUE)}
}
\value{
list of parameters
}
\description{
Function to set root_water_uptake while maintaining unused default model parameters.
}
\details{
\code{root_water_uptake_model}  options:
\describe{
\item{\code{root_water_uptake_model = 0}}{Feddes model. Column names of \code{root_water_uptake_params}: \code{"critical_stress_index", "P0", "POpt", "P2H", "P2L", "P3", "r2H", "r2L"}}
\item{\code{root_water_uptake_model = 1}}{S-Shaped model. Column names of \code{root_water_uptake_params}: \code{"critical_stress_index", "h50", "P3"}}
\item{\code{root_water_uptake_model = 2}}{Nimah \& Hanks model. Column names of \code{root_water_uptake_params}: \code{"lHydRed", "P3"}}
\item{\code{root_water_uptake_model = 3}}{Couvreur model. Column names of \code{root_water_uptake_params}: \code{"hx_min", "rKrs", "rKComp"}}
\item{\code{root_water_uptake_model = 4}}{de Jong van Lier model. Column names of \code{root_water_uptake_params}: \code{"RootRad", "XylemRad", "Conductance", "RootCond", "a", "H-Wilt"}}
}
}
\examples{
hydrus_model$root_water_uptake <- root_water_uptake(root_water_uptake_model = 1,
                                                              root_water_uptake_params = data.frame(
                                                              h50 = -800,
                                                              P3 = 3
                                                              ))
}
