% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizing_functions.R
\name{soil_hydraulics}
\alias{soil_hydraulics}
\title{Set soil hydraulic parameters in a HYDRUS model}
\usage{
soil_hydraulics(
 soil_hydraulic_model = 0,
 mobile_immobile = FALSE,
 hysteresis = 0,
 initially_drying_curve = TRUE,
 soil_hydraulic_parameters = data.frame(material = 1,
                                        theta_r = 0.045,
                                        theta_s = 0.43,
                                        alpha = 0.145,
                                        n = 2.68,
                                        K_s = 712.8,
                                        l = 0.5)
)
}
\arguments{
\item{soil_hydraulic_model}{integer indicating which soil hydraulic model to use. See details.}

\item{mobile_immobile}{logical.}

\item{hysteresis}{integer indicating which hysteresis model to use. Default = 0; no hysteresis. See details.}

\item{initially_drying_curve}{logical. Only used when \code{hysteresis = TRUE}. Initially drying curve = TRUE; initially wetting curve = FALSE.}

\item{soil_hydraulic_parameters}{data.frame of soil hydraulics parameters. Column names must match the specified soil hydraulic model and number of rows must match the number of materials in the model. See details for parameters.}

\item{surface_flow_into_fracture}{numeric fraction in [0,1]. Used when \code{soil_hydraulic_model = 8}; the dual-permeability model.}
}
\value{
list of parameters
}
\description{
Function to set soil_hydraulics while maintaining unused default model parameters.
}
\details{
\code{soil_hydraulic_model} options:
\describe{
  \item{\code{soil_hydraulic_model = 0}}{van Genuchten-Maulem single-porosity model (van Genuchten 1980). \code{soil_hydraulic_parameters}: \code{"material", "theta_r", "theta_s", "alpha", "n", "K_s", "l"}}
  \item{\code{soil_hydraulic_model = 1}}{Modified van Genuchten single-porosity model (Vogel and Cislerova, 1988). \code{soil_hydraulic_parameters}: \code{"material", "theta_r", "theta_s", "alpha", "n", "K_s", "l", "theta_m", "theta_a", "theta_k","K_k"}}
  \item{\code{soil_hydraulic_model = 2}}{Brooks-Corey (1964) single-porosity model. \code{soil_hydraulic_parameters}: \code{"material", "theta_r", "theta_s", "alpha", "n", "K_s", "l"}}
  \item{\code{soil_hydraulic_model = 3}}{van Genuchten-Maulem with air-entry value of -2cm. \code{soil_hydraulic_parameters}: \code{"material", "theta_r", "theta_s", "alpha", "n", "K_s", "l"}}
  \item{\code{soil_hydraulic_model = 4}}{Kosugi (1996) single-porosity (the log-normal model). \code{soil_hydraulic_parameters}: \code{"material", "theta_r", "theta_s", "alpha", "n", "K_s", "l"}}
  \item{\code{soil_hydraulic_model = -4}}{Kosugi single-porosity model with the Brunswick modification (Weber et al., 2019). \code{soil_hydraulic_parameters}: \code{"material", "theta_r", "theta_s", "alpha", "n", "K_s", "l", "K_s_ncapt", "a_film", "pf0", "hr"}}
  \item{\code{soil_hydraulic_model = 5}}{Dual-porosity (Durner, 1994; the dual van Genuchten-Maulem model). \code{soil_hydraulic_parameters}: \code{"material", "theta_r", "theta_s", "alpha", "n", "K_s", "l", "w_2", "alpha_2", "n_2".}}
  \item{\code{soil_hydraulic_model = 6}}{Dual-porosity nonequilibrium flow model (mobile-immobile water, water content mass transfer). \code{soil_hydraulic_parameters}: \code{"material", "theta_r", "theta_s", "alpha", "n", "K_s", "l", "theta_r_im", "theta_s_im", "omega".}}
  \item{\code{soil_hydraulic_model = 7}}{Dual-porosity nonequilibrium flow model (mobile-immobile water, pressure head mass transfer). \code{soil_hydraulic_parameters}: \code{"material", "theta_r", "theta_s", "alpha", "n", "K_s", "l", "theta_r_im", "theta_s_im", "alpha_im", "n_im", "omega".}}
  \item{\code{soil_hydraulic_model = 8}}{Dual-permeability nonequilibrium flow model. \code{soil_hydraulic_parameters}: \code{"material", "theta_r", "theta_s", "alpha", "n", "K_s", "l", "theta_r_fr", "theta_s_fr", "alpha_fr", "n_fr", "K_s_fr", "l_fr", "w", "beta", "gamma", "a", "K_a".}}
  \item{\code{soil_hydraulic_model = 9}}{Look-up Tables. \code{soil_hydraulic_parameters}: \code{"material", "theta_r", "theta_s", "K_s"}}
}
\code{hysteresis} options:
\describe{
 \item{\code{hysteresis = 0}}{No hysteresis}
 \item{\code{hysteresis = 1}}{Hysteresis in retention curve (Kool \& Parker, 1987)}
 \item{\code{hysteresis = 2}}{Hysteresis in retention curve and conductivity}
 \item{\code{hysteresis = 3}}{Hysteresis in retention curve (no pumping; Lenhard and Parker, 1992)}
}
}
\examples{
hydrus_model$soil_hydraulics <- soil_hydraulics(soil_hydraulic_model = 8,
                                                          soil_hydraulic_parameters =
                                                          data.frame(material = 1:3,
                                                                     theta_r = c(0.0792, 0.0813, 0.05),
                                                                     theta_s = c(0.47, 0.4231, 0.37),
                                                                     alpha = c(0.03, 0.019, 0.0353),
                                                                     n = c(1.35, 1.45, 3.18),
                                                                     K_s = c(6, 8, 642.98),
                                                                     l = 0.5,
                                                                     theta_r_fr = c(0.01298, 0, 0.01),
                                                                     theta_s_fr = c(0.43903, 0.3, 0.8),
                                                                     alpha_fr = c(0.017304, 0.079, 0.01),
                                                                     n_fr = c(1.3202, 1.6, 1.01),
                                                                     K_s_fr = c(289.89, 75.5, 386.52),
                                                                     l_fr = 0.5,
                                                                     w = 0.04,
                                                                     beta = c(4.74, 2, 1),
                                                                     gamma = c(0.01, 0.4, 0.4),
                                                                     a = c(2.197, 7.8, 10),
                                                                     K_a = c(1.4143, 0.01, 1.497))
}
