% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizing_functions.R
\name{solute_options}
\alias{solute_options}
\title{Set solute parameters in a HYDRUS model}
\usage{
solute_options(
 standard_solute_transport = TRUE,
 number_solutes = 0,
 equilibrium_adsorption = FALSE
)
}
\arguments{
\item{standard_solute_transport}{logical. Use standard solute transport (TRUE)}

\item{number_solutes}{integer. Number of solutes.}

\item{equilibrium_adsorption}{logical. Use equilibrium adsorption (TRUE)}
}
\value{
list of parameters
}
\description{
Function to set solute_options while maintaining unused default model parameters.
}
\examples{
hydrus_model$solute_options <- solute_options(number_solutes = 1)
}
