% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizing_functions.R
\name{solute_transport}
\alias{solute_transport}
\title{Set solute trasnport parameters in a HYDRUS model}
\usage{
solute_transport(
 standard_solute_transport = TRUE,
 number_solutes = 0,
 equilibrium_adsorption = FALSE
)
}
\arguments{
\item{solute_transport_bcs}{}
}
\value{
list of parameters
}
\description{
Function to set solute_transport while maintaining unused default model parameters.
}
\examples{
hydrus_model$solute_options <- solute_options(number_solutes = 1)
}
