% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizing_functions.R
\name{time_parameters}
\alias{time_parameters}
\title{Set time parameters in a HYDRUS model}
\usage{
time_parameters(
 initial_time_step = 0.001,
 minimum_time_step = 1e-05,
 maximum_time_step = 5,
 lower_time_step_mult = 1.3,
 upper_time_step_mult = 0.7,
 lower_optim_iter_range = 3,
 upper_optim_iter_range = 7,
 initial_model_time = 0,
 final_model_time = 100
)
}
\arguments{
\item{initial_time_step}{numeric. initial time step.}

\item{minimum_time_step}{numeric.}

\item{maximum_time_step}{numeric.}

\item{lower_time_step_mult}{numeric.}

\item{upper_time_step_mult}{numeric.}

\item{lower_optim_iter_range}{numeric.}

\item{upper_optim_iter_range}{numeric.}

\item{initial_model_time}{integer. Model time 1 in time_unit of model.}

\item{final_model_time}{integer. Last model time in time_unit of model.}
}
\value{
list of parameters
}
\description{
Function to set time_parameters while maintaining unused default model parameters.
}
\examples{
hydrus_model$time_parameters <- time_parameters(final_model_time = 365)
}
