% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizing_functions.R
\name{time_variable_bc}
\alias{time_variable_bc}
\title{Set time variable boundary conditions parameters in a HYDRUS model}
\usage{
time_variable_bc(
 time_variable_bc = FALSE,
 meteorological_data = FALSE,
 repeat_bc_records = FALSE,
 repeat_bc_records_n = 1,
 daily_var_transpiration = FALSE,
 sinusoidal_var_precipitation = FALSE
)
}
\arguments{
\item{time_variable_bc}{logical. Use time variable boundary condition option (TRUE) or not (FALSE)}

\item{meteorological_data}{logical. Use meteorological data option (TRUE) or not (FALSE)}

\item{repeat_bc_records}{logical. Use repeat boundary conditions option (TRUE) or not (FALSE)}

\item{repeat_bc_records_n}{integer. Number of times to repeat boundary condition records}

\item{daily_var_transpiration}{logical. Option to use daily variations of transpiration during day generated by HYDRUS}

\item{sinusoidal_var_precipitation}{logical. Option to use sinusoidal variations of precipitation generated by HYDRUS}
}
\value{
list pf parameters
}
\description{
Function to set time_variable_bc while maintaining unused default model parameters.
}
\examples{
hydrus_model$time_variable_bc <- time_variable_bc(time_variable_bc = TRUE,
                                                            meteorological_data = TRUE)
}
