% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizing_functions.R
\name{water_flow_bcs}
\alias{water_flow_bcs}
\title{Set water flow boundary conditions in a HYDRUS model}
\usage{
water_flow_bcs(
 upper_bc = "constant_pressure_head",
 initial_condition = FALSE,
 lower_bc = "free_drainage",
 triggered_irrigation = FALSE,
 constant_boundary_fluxes = data.frame(upper_bc_flux = 0,
                                       lower_bc_flux = 0,
                                       root_water_uptake_flux = 0),
 irrigation_params = data.frame(trigger_node = 1,
                                trigger_pressure_head = -100,
                                irrigation_rate = 0,
                                irrigation_duration = 0,
                                lag_time = 0)
)
}
\arguments{
\item{upper_bc}{character string of the upper boundary condition. See details for options.}

\item{initial_condition}{logical.}

\item{lower_bc}{character string of the lower boundary condition. See details for options.}

\item{triggered_irrigation}{logical. Option to use triggered irrigation (TRUE), only available with \code{upper_bc = "atm_bc_with_surface_layer"} or \code{upper_bc = "atm_bc_with_surface_runoff"}.}

\item{constant_boundary_fluxes}{data.frame of fluxes when using the \code{upper_bc = "constant_boundary_flux"} option}

\item{irrigation_params}{data.frame of triggered irrigation parameters}
}
\value{
list of parameters
}
\description{
Function to set water_flow_bcs while maintaining unused default model parameters.
}
\details{
\code{upper_bc} options:
\describe{
 \item{\code{upper_bc = "constant_pressure_head"}}{}
 \item{\code{upper_bc = "constant_flux"}}{Need \code{constant_boundary_fluxes} data.frame with column names \code{"upper_bc_flux", "lower_bc_flux", "root_water_uptake_flux"}.}
 \item{\code{upper_bc = "atm_bc_with_surface_layer"}}{When \code{triggered_irrigation = TRUE}, column names of \code{irrigation_params} must be: "trigger_node", "trigger_pressure_head", "irrigation_rate", "irrigation_duration", "lag_time".}
 \item{\code{upper_bc = "atm_bc_with_surface_runoff"}}{When \code{triggered_irrigation = TRUE}, column names of \code{irrigation_params} must be: "trigger_node", "trigger_pressure_head", "irrigation_rate", "irrigation_duration", "lag_time".}
 \item{\code{upper_bc = "variable_pressure_head"}}{}
 \item{\code{upper_bc = "variable_pressure_head/flux"}}{}
}
\code{lower_bc} options:
\itemize{
 \item{\code{lower_bc = "constant_pressure_head"}}
 \item{\code{lower_bc = "constant_flux"}}
 \item{\code{lower_bc = "variable_pressure_head"}}
 \item{\code{lower_bc = "variable_flux"}}
 \item{\code{lower_bc = "free_drainage"}}
 \item{\code{lower_bc = "deep_drainage"}}
 \item{\code{lower_bc = "seepage_face"}}
 \item{\code{lower_bc = "horizontal_drains"}}
}
}
\examples{
hydrus_model$water_flow_bcs <- water_flow_bcs(upper_bc = "constant_flux")
}
