% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_c_time_info.R
\name{block_c_time_info}
\alias{block_c_time_info}
\title{Set Time Parameters in BLOCK C of SELECTOR.IN}
\usage{
block_c_time_info(
  hydrus_project,
  time_parameters = time_parameters,
  print_options = print_options
)
}
\arguments{
\item{hydrus_project}{a hydrus project object created with
'create_hydrus_project'}

\item{time_parameters}{list of time parameters, see details.}

\item{print_options}{list of print options, see details.}
}
\description{
desc
}
\details{
Time parameters are all numeric and in the 'time_unit' of the 'hydrus_model'.
Time parameters include 'initial_time_step', 'minimum time step', 'maximum
time step', 'lower_time_step_mult', 'upper_time_step_mult',
'lower_optim_iter_range', 'upper_optim_iter_range', 'initial model time',
'final_model_time'. Print options include 'print_times', a data.frame with one column of 'times' to
print out. 'interval_output_option' is a logical indicating to turn on interval output
(default is off, FALSE). 'interval_output' is used when
'interval_output_option = TRUE' and specfies the interval to output.
'print_time_interval' is a logical indicating whether or not to print time
information (default on, TRUE). 'time_info_print_every_n_time_steps', default 1, print time interval.
}
