% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_d_root_growth_info.R
\name{block_d_root_growth_info}
\alias{block_d_root_growth_info}
\title{Set Block D of SELECTOR.IN: Root Growth Information}
\usage{
block_d_root_growth_info(hydrus_project, root_growth = root_growth)
}
\arguments{
\item{hydrus_project}{a hydrus project object created with
'create_hydrus_project'}

\item{root_growth_depth}{how root growth should be specified. See details for options. Default '0', root depth specified using time-variable boundary conditions.}

\item{root_growth_params}{data.frame of parameters for the root growth specification. Default NULL because no additional parameters when root_growth_depth = 0.}

\item{root_growth_factor}{either '0' for root growth factor from given data or '1' for a root growth factor of 50% after 50% growing season.}
}
\description{
Set Block D of SELECTOR.IN: Root Growth Information
}
