% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_g_root_water_uptake_info.R
\name{block_g_root_water_uptake_info}
\alias{block_g_root_water_uptake_info}
\title{Set Block G of SELECTOR.IN: Root Water Uptake Information}
\usage{
block_g_root_water_uptake_info(
  hydrus_project,
  root_water_uptake = root_water_uptake
)
}
\arguments{
\item{hydrus_project}{a hydrus model object created with
'create_hydrus_project'}

\item{root_water_uptake_model}{which root water uptake model to use. Default is '0' the Feddes model.}

\item{root_water_uptake_params}{data.frame of parameters (by column) for the root water uptake model.}
}
\description{
Set Block G of SELECTOR.IN: Root Water Uptake Information
}
