% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_hydrus_project.R
\name{create_hydrus_project}
\alias{create_hydrus_project}
\title{Initiates a HYDRUS project}
\usage{
create_hydrus_project(
  project_name,
  parent_dir,
  hydrus_version = 5,
  description = NULL
)
}
\arguments{
\item{project_name}{Name of the HYDRUS project to create}

\item{parent_dir}{Path to the project folder}

\item{hydrus_version}{Integer indicating which HYDRUS version to use. Must be version 4 or 5.}

\item{description}{Optional project description character string}
}
\value{
an object of hydrus_project class with all model info set
}
\description{
Initiates a HYDRUS project
}
