% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_dat_file.R
\name{edit_dat_file}
\alias{edit_dat_file}
\title{Edits the HYDRUS1D.DAT file}
\usage{
edit_dat_file(hydrus_model)
}
\arguments{
\item{hydrus_model}{a hydrus project object created with
'create_hydrus_project'}
}
\value{
Edits the .DAT file of the HYRDUS project
}
\description{
Edits the HYDRUS1D.DAT file
}
\examples{
edit_dat_file(hydrus_model)
}
