% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_data_file.R
\name{edit_dat_file}
\alias{edit_dat_file}
\title{Title}
\usage{
edit_dat_file(
  hydrus_project,
  main_processes = main_processes,
  units = units,
  geometry = geometry,
  print_options = print_options,
  water_flow_bcs = water_flow_bcs
)
}
\arguments{
\item{water_flow_bcs}{}
}
\description{
Title
}
