% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_soil_profile.R
\name{edit_soil_profile}
\alias{edit_soil_profile}
\title{Set PROFILE.DAT file}
\usage{
edit_soil_profile(
  hydrus_model,
  mesh_density = data.frame(fixed_points = c(0, -101), upper_relative_size_fe = 1,
    lower_relative_size_fe = 1),
  set_mesh_nodes_manually = FALSE,
  mesh_nodes_manual = data.frame(number = 1:101, z = -0:-100),
  nodal_soil_properties = list(h = -101, root = 0, a_xz = 1, b_xz = 1, d_xz = 1, mat = 1,
    lay = 1, temp = 20, conc = 0)
)
}
\arguments{
\item{nodal_soil_properties}{}
}
\value{
edits PROFILE.DAT file
}
\description{
Set PROFILE.DAT file
}
\examples{
edit_soil_profile(hydrus_model)
}
