% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterize_hydrus_model.R
\name{parameterize_hydrus_model}
\alias{parameterize_hydrus_model}
\title{Title}
\usage{
parameterize_hydrus_model(
  hydrus_project,
  main_processes = list(water_flow = TRUE, vapor_flow = FALSE, snow_hydrology = FALSE,
    particle_tracking = FALSE, solute_transport = FALSE, unsatchem = FALSE, HP1 = FALSE,
    heat_transport = FALSE, root_water_uptake = FALSE, root_growth = FALSE, inverse =
    FALSE),
  units = list(time_unit = "days", space_unit = "cm", mass_unit = "mmol"),
  geometry = list(number_materials = 1, number_subregions = 0, number_nodes = 101,
    profile_depth = 100, observation_nodes = NULL),
  print_options = list(print_times = TRUE, screen_output = TRUE, print_fluxes_not_temp =
    TRUE, times_to_print = data.frame(times = c(100)), interval_output_option = FALSE,
    interval_output = 0.001, print_time_information = TRUE,
    time_info_print_every_n_time_steps = 1),
  time_parameters = list(initial_time_step = 0.001, minimum_time_step = 1e-05,
    maximum_time_step = 5, lower_time_step_mult = 1.3, upper_time_step_mult = 0.7,
    lower_optim_iter_range = 3, upper_optim_iter_range = 7, initial_model_time = 0,
    final_model_time = 100),
  time_variable_bc = list(time_variable_bc = FALSE, meteorological_data = FALSE),
  iteration_criteria = list(maximum_iterations = 10, water_content_tol = 0.001,
    pressure_head_tol = 1, internal_interpolation_tables = TRUE,
    lower_limit_tension_interval = 1e-06, upper_limit_tension_interval = 10000),
  soil_hydraulics = c(soil_hydraulic_model = 0, mobile_immobile = FALSE, hysteresis = 0,
    initially_drying_curve = TRUE, soil_hydraulic_parameters = data.frame(material = 1,
    theta_r = 0.045, theta_s = 0.43, alpha = 0.145, n = 2.68, K_s = 712.8, l = 0.5)),
  water_flow_bcs = list(upper_bc = "constant_pressure_head", initial_condition = FALSE,
    lower_bc = "free_drainage", triggered_irrigation = FALSE),
  root_water_uptake = list(root_water_uptake_model = 0, root_water_uptake_params =
    data.frame(critical_stress_index = 1, P0 = -10, POpt = -25, P2H = -200, P2L = -800,
    P3 = -8000, r2H = 0.5, r2L = 0.1)),
  root_growth = list(root_growth_depth = 0, root_growth_params = NULL, root_growth_factor
    = NULL),
  solute_options = list(number_solutes = 0, equilibrium_adsorption = FALSE),
  particle_tracking = list(init_water_storage = 2, cumulative_surface_flux = 2)
)
}
\arguments{
\item{water_flow_bcs}{}
}
\description{
Title
}
