% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_hydrus_1D.R
\name{run_hydrus_1D}
\alias{run_hydrus_1D}
\title{Executes HYDRUS 1D executable}
\usage{
run_hydrus_1D(hydrus_model, hydrus_path = NULL, show_output = TRUE)
}
\arguments{
\item{hydrus_model}{hydrus model}

\item{hydrus_path}{Path to the Hydrus 1D executable. NULL will use default install path used by HYDRUS.}

\item{show_output}{Logical (Default = TRUE) whether the shell
output should be visible on R console (relevant only on windows)}
}
\description{
Executes HYDRUS 1D executable
}
